third_party/blink/web_tests/fast/dom/Window/window-postmessage-clone-really-deep-array-expected.txt

Tests that we abort cloning overdeep arrays.
PASS: 'postMessage(reallyDeepArray)' threw RangeError: Maximum call stack size exceeded
PASS: eventData is done of type string
